I have created a new folders controller, service, interface, service. And I have moved these files there.
